api-gateway/requirements.txt
flask
requests
grpcio
protobuf
googleapis-common-protos
PyJWT
marshmallow
kafka-python